put each benchmark in its own file
